feat: adicionando features da branch do joao pra essa
