+++ .gitignore
@@ -1,3 +1,7 @@
 /bin
 
 /obj
+
+eda_2_p1.depend
+
+eda_2_p1.layout
